% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_CI_cols}
\alias{add_CI_cols}
\title{Adds columns for CI}
\usage{
add_CI_cols(df)
}
\arguments{
\item{df}{data.frame with point estimates}
}
\value{
data.frame with two additional columns for lower and upper CI containing NAs
}
\description{
Adds columns for CI
}
\keyword{internal}
