% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass}
\alias{biomass}
\title{biomass dataset}
\description{
biomass dataset
}
\details{
This is an simulated dataset about grassland biodiversity and biomass.

In brief the imaginary sampling design of the simulated dataset is as follows.
Invertebrates were sampled once every year over 10 successive years (`Year`) from 20
different populations (`Population`). For each sample, the temperature (`Temperature`)
and (`Precipitation`) were measured and overall species diversity (`SpeciesDiversity`)
and biomass were recorded (`Biomass`).
}
\keyword{datasets}
