% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boot_to_df}
\alias{boot_to_df}
\title{List to data.frame with bootstrap samples per row}
\usage{
boot_to_df(lcol)
}
\arguments{
\item{lcol}{A bootstrap list-column}
}
\value{
data.frame with bootstrap estimates as list-column
}
\description{
List to data.frame with bootstrap samples per row
}
\keyword{internal}
