% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{calc_CI}
\alias{calc_CI}
\title{Calculates CI from bootstrap replicates}
\usage{
calc_CI(x, CI)
}
\arguments{
\item{x}{numeric vector}

\item{CI}{CI level, e.g. 0.95}
}
\description{
Calculates CI from bootstrap replicates
}
\keyword{internal}
