% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{calc_var_r}
\alias{calc_var_r}
\title{Calculates Distribution specific and Residual variance}
\usage{
calc_var_r(mod, expect = "meanobs")
}
\arguments{
\item{mod}{fitted model}

\item{expect}{expectation in poisson and binomial models, atm only "meanobs"}
}
\description{
Calculates Distribution specific and Residual variance
}
\keyword{internal}
