% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{fixef_simple}
\alias{fixef_simple}
\title{Get tidy summary of fixed effect estimates}
\usage{
fixef_simple(mod, intcp = FALSE)
}
\arguments{
\item{mod}{merMod object}

\item{intcp}{include intercept in summary? Defaults to FALSE}
}
\value{
tibble with two columns: term and estimate
}
\description{
Get tidy summary of fixed effect estimates
}
\keyword{internal}
