% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_mods.R
\name{get_bw_scaling}
\alias{get_bw_scaling}
\title{Standardise variables}
\usage{
get_bw_scaling(mod)
}
\arguments{
\item{mod}{merMod object.}
}
\value{
list with scaling factors to multiply estimates with to get BW
}
\description{
Standardise variables
}
\keyword{internal}
