% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{get_ndf}
\alias{get_ndf}
\title{Get numerator dfs for reduced models}
\usage{
get_ndf(partvar, mod, dat)
}
\arguments{
\item{partvar}{One or more fixed effect variables which are taken out
of the model.}

\item{mod}{merMod object.}

\item{dat}{Data.frame to fit the model}
}
\value{
Numeric vector, giving numerator degrees of freedom
}
\description{
Get numerator dfs for reduced models
}
\keyword{internal}
