% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_comps.R
\name{get_var_comps}
\alias{get_var_comps}
\title{Extract variance components from merMod.}
\usage{
get_var_comps(mod, expct)
}
\arguments{
\item{mod}{A merMod object.}

\item{expct}{expectation.}
}
\value{
Fixed, random and residual variance
}
\description{
Extract variance components from merMod.
}
\keyword{internal}
