% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Grasshoppers}
\alias{Grasshoppers}
\title{Grasshoppers dataset}
\description{
Grasshoppers dataset
}
\details{
This is a real dataset from grasshoppers.

This dataset contains data on spatial variation in color morph ratios in a
color-polymorphic species of grasshopper (Dieker et al 2018). Individuals of this species occur
either in green or a brown color variant and the dataset contains counts of
brown and green individuals (seprarated for females and males) from 42 sites
sampled in the field. All 'Bio' variables describe various aspects of ecologically
relevant climatic conditions (see Karger et al. 2017).
}
\references{
Dieker, P., L. Beckmann, J. Teckentrup, and H. Schielzeth (2018)
\emph{Spatial analyses of two colour polymorphisms in an alpine grasshopper reveal
a role of small-scale heterogeneity}. Ecology and Evolution, 8, 7273-7284.

Karger, D. N., O. Conrad, J. Bohner, T. Kawohl, H. Kreft, R. W. Soria-Auza,
N. E. Zimmermann, H. P. Linder, and M. Kessler (2017) \emph{Data descriptor: Climatologies
at high resolution for the earth's land surface areas}. Scientific Data, 4, 170122.
}
\keyword{datasets}
