% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{group_vars}
\alias{group_vars}
\title{Calculates / extracts variance components from random effects and random slopes}
\usage{
group_vars(grname, var_comps, mod)
}
\arguments{
\item{grname}{The name of a grouping factor, usually accessed by looping over the
grname argument of the rptR functions.}

\item{var_comps}{A list. Output of the lme4::VarCorr function.}

\item{mod}{An lme4 model object.}
}
\description{
This function uses the method from Paul Johnson to compute the average group
variance across the levels of a covariate.
}
\keyword{internal}
