% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{GuineaPigs}
\alias{GuineaPigs}
\title{GuineaPigs dataset}
\description{
GuineaPigs dataset
}
\details{
This is a real dataset from guinea pigs.

The dataset contains testosterone measurements (`Testo`) of 31 male guinea pigs, each measured at 5 time points.
(age between 120 and 240 days at 30-day intervals). As covariates the dataset
contains the time point of measurement (`Time`) and a rank index derived from behavioral
observations (`Rank`) around the time of measurement (see Mutwill et al. in prep. for details).
}
\keyword{datasets}
