% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{part_R2s}
\alias{part_R2s}
\title{Calculate part R2}
\usage{
part_R2s(
  mod,
  expct,
  overdisp_name,
  R2_type,
  all_comb,
  partition,
  data_mod,
  allow_neg_r2
)
}
\arguments{
\item{mod}{merMod object}

\item{expct}{Expectation}

\item{overdisp_name}{Name of overdispersion term}
}
\value{
R2, atm data.frame with one element
}
\description{
Calculate part R2
}
\examples{

}
