% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partR2-package.R
\docType{package}
\name{partR2-package}
\alias{partR2-package}
\alias{_PACKAGE}
\title{partR2: partitioning R2 in mixed models}
\description{
partR2 provides R2, partitioned R2s, structure coefficients and
estimates for for mixed models.
}
\details{
It's main goals are:

\itemize{
\item Estimate marginal and conditional R2 for LMMs and GLMMs.
\item Partition the R2 into variance explained uniquely by each predictor
and variance explained by a combination of predictors.
\item Provide structure coefficients (the correlation between each
predictor and the predicted response, independent of the other predictors)
\item Report model estimates (based on the broom.mixed package)
\item Use parametric bootstrapping to get confidence intervals for
all estimates.
}
}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
