% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partR2-package.R, R/partR2.R
\docType{package}
\name{partR2}
\alias{partR2}
\alias{partR2-package}
\title{\code{partR2} package
R2, commonality coefficients and structure coefficients for mixed models}
\usage{
partR2(mod, partvars = NULL, R2_type = "marginal", cc_level = NULL,
  nboot = NULL, CI = 0.95, parallel = FALSE, ncores = NULL,
  expct = "meanobs")
}
\arguments{
\item{mod}{merMod object fitted with lme4.}

\item{partvars}{Character vector specifying the predictors for which to partition the R2.}

\item{R2_type}{"marginal" or "conditional" R2.}

\item{cc_level}{Level up to which commonality coefficients are calculated.
The number of sets for which to calculate partial R2 increases exponantially,
i.e. for 10 variables 2^10 - 1 commonality coefficients and potentially CIs
can be calculated. To limit this number for models with many fixed effects,
specify cc_level. cc_level = 3 would limit the number of sets to a maximum of
triplets, i.e. calculating partial R2 for each individual fixed effect specified in
partvars, for each pair and for each triplet.}

\item{nboot}{Number of parametric bootstraps for interval estimation
(defaults to NULL). Larger numbers of bootstraps give a better
asymtotic CI, but may be time-consuming. Bootstrapping can be switch on by setting
\code{nboot = 1000}.}

\item{CI}{Width of the required confidence interval between 0 and 1 (defaults to
0.95).}

\item{parallel}{If TRUE, computation runs in parallel, leaving one CPU free, except ncores is specified.}

\item{ncores}{number of cpus for parallel computation}

\item{expct}{A string specifying the method for estimating the expectation in Poisson models
with log link and in Binomial models with logit link (in all other cases the agrument is ignored).
The only valid terms are 'meanobs' and 'latent' (and 'liability for binary and proportion data).
With the default 'meanobs', the expectation is estimated as the mean of the observations in the sample.
With 'latent', the expectation is estimated from estiamtes of the intercept and variances on the link scale.
While this is a preferred solution, it is susceptible to the distribution of fixed effect covariates and gives
appropriate results typically only when all covariances are centered to zero. With 'liability'
estimates follow formulae as presented in Nakagawa & Schielzeth (2010).}
}
\value{
Returns an object of class \code{partR2} that is a a list with the following elements:
\item{call}{model call}
\item{R2_type}{Marginal or conditional R2}
\item{R2_pe_ci}{R2 and confidence intervals for full model and partitions}
\item{SC_pe_ci}{Structure coefficients and confidence intervals}
\item{Ests_pe_ci}{Model estimates and confidence intervals. Point estimates
were extracted with broom.mixed::tidy}
\item{R2_boot}{Parametric bootstrap samples for R2 for full model and partitions}
\item{SC_boot}{Parametric bootstrap samples for structure coefficients}
\item{Ests_pe_ci}{Parametric bootstrap samples for model estimates}
\item{partvars}{predictors to partition}
\item{CI}{Coverage of the confidence interval as specified by the \code{CI} argument.}
\item{boot_warnings}{Potential warnings from estimating partial R2s during
parametric bootstrapping}
\item{boot_message}{Potential messages from estimating partial R2s
during parametric bootstrapping. Common are for example singularity messages
in lme4.}
}
\description{
R2, commonality coefficients and structure coefficients for gaussian lme4 models.
}
\examples{

data(biomass)
library(lme4)

# Gaussian data
mod <- lmer(Biomass ~  Year + Temperature * Precipitation + SpeciesDiversity + (1|Population),
            data = biomass)
# Only R2 with CI
(R2 <- partR2(mod, R2_type = "marginal", nboot = 15, CI = 0.95))

# Partitioned R2
(R2 <- partR2(mod,  partvars = c("SpeciesDiversity", "Temperature:Precipitation",
                                 "Temperature", "Precipitation"),
                                 R2_type = "marginal", nboot = 10, CI = 0.95))



}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.

Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
