% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.partR2.R
\name{plot.partR2}
\alias{plot.partR2}
\title{Plot a partR2 object}
\usage{
\method{plot}{partR2}(x, toplot)
}
\arguments{
\item{x}{A partR2 object.}

\item{toplot}{Commonality coefficients ("CC") or structure coefficients ("SC")}
}
\description{
Forestplot of the partR2 results
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com),
        Holger Schielzeth  (holger.schielzeth@uni-jena.de)
        Shinichi Nakagawa (s.nakagawa@unsw.edu.au),
}
\keyword{models}
