% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.partR2.R
\name{print.partR2}
\alias{print.partR2}
\title{Print a partR2 object}
\usage{
\method{print}{partR2}(x, ...)
}
\arguments{
\item{x}{An partR2 object returned from one of the partR2 functions}

\item{\dots}{Additional arguments; none are used in this method.}
}
\value{
Abbreviations in the print.rpt output:
\item{R}{Repeatability.}
\item{SE}{Standard error of R.}
\item{CI}{Confidence interval of R derived from parametric bootstrapping.}
\item{P}{P-value}
\item{LRT}{Likelihood-ratio test}
\item{Permutation}{Permutation of residuals}
}
\description{
Displays the results a partR2object (i.e. the result of a partR2 function call) in a nice form.
}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com),
        Shinichi Nakagawa (s.nakagawa@unsw.edu.au),
        Holger Schielzeth  (holger.schielzeth@uni-jena.de)
}
\keyword{models}
