% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{R2_of_red_mod}
\alias{R2_of_red_mod}
\title{Calculate R2 from a reduced model}
\usage{
R2_of_red_mod(partvar, mod, R2_pe, dat, expct, overdisp_name, R2_type)
}
\arguments{
\item{partvar}{One or more fixed effect variables which are taken out
of the model.}

\item{mod}{merMod object.}

\item{R2_pe}{R2 function.}

\item{expct}{Expectation}

\item{overdisp_name}{Name of overdispersion term}

\item{R2_type}{"marginal" or "conditional"}

\item{data}{Data.frame to fit the model}
}
\value{
R2 of reduced model.
}
\description{
Calculate R2 from a reduced model
}
\keyword{internal}
