% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{R2_pe}
\alias{R2_pe}
\title{Calculate R2}
\usage{
R2_pe(mod, expct, overdisp_name)
}
\arguments{
\item{mod}{merMod object}

\item{expct}{Expectation}

\item{overdisp_name}{Name of overdispersion term}
}
\value{
R2, atm data.frame with one element
}
\description{
Calculate R2
}
\examples{

}
