% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{SC_pe}
\alias{SC_pe}
\title{Structure coefficients}
\usage{
SC_pe(mod)
}
\arguments{
\item{mod}{merMod object.}
}
\value{
data.frame with structure coefficients
}
\description{
Structure coefficients
}
\keyword{internal}
