% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sim_data}
\alias{sim_data}
\title{sim_data dataset}
\description{
sim_data dataset
}
\details{
Simulated dataset to test random slope models. Source:
https://stats.stackexchange.com/questions/162553/how-to-simulate-a-random-slope-model
}
\keyword{internal}
