% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_comps.R
\name{var_comps_binary}
\alias{var_comps_binary}
\title{Get variance components for binomial model with binary response.}
\usage{
var_comps_binary(mod, expct)
}
\arguments{
\item{mod}{merMod object with binomial family and binary response.}

\item{expct}{"latent", "meanobs" of "liability"}
}
\value{
Fixed, random and residual variance
}
\description{
Get variance components for binomial model with binary response.
}
\keyword{internal}
