% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_comps.R
\name{var_comps_gaussian}
\alias{var_comps_gaussian}
\title{Get variance components for gaussian model.}
\usage{
var_comps_gaussian(mod, ...)
}
\arguments{
\item{mod}{merMod object with gaussian family.}
}
\value{
Fixed, random and residual variance
}
\description{
Get variance components for gaussian model.
}
\keyword{internal}
