% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{var_comps_proportion}
\alias{var_comps_proportion}
\title{Get variance components for binomial model with proportion response.}
\usage{
var_comps_proportion(mod, expct, overdisp_name)
}
\arguments{
\item{mod}{merMod object with binomial family and binary response.}

\item{expct}{"latent", "meanobs" of "liability"}
}
\value{
Fixed, random and residual variance
}
\description{
Get variance components for binomial model with proportion response.
}
\keyword{internal}
