% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{with_warnings}
\alias{with_warnings}
\title{Captures and suppresses (still to find out why) warnings of an expression}
\usage{
with_warnings(expr)
}
\arguments{
\item{expr}{An expression, such as the sequence of code used by rptR to calculate
bootstrap or permutation estimates}
}
\description{
This function is used within rptR to capture lme4 model fitting warnings in the
bootstrap and permutation procedures.
}
\keyword{internal}
