% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.partR2.R
\name{forestplot.partR2}
\alias{forestplot.partR2}
\title{Plot a partR2 object}
\usage{
forestplot.partR2(x)
}
\arguments{
\item{x}{A partR2 object.}
}
\description{
Forestplot of the partR2 results
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com),
        Holger Schielzeth  (holger.schielzeth@uni-jena.de)
        Shinichi Nakagawa (s.nakagawa@unsw.edu.au),
}
\keyword{models}
