% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.partR2.R
\name{forestplot}
\alias{forestplot}
\title{Plot a partR2 object}
\usage{
forestplot(x, type = c("R2", "Ests", "SC"))
}
\arguments{
\item{x}{A partR2 object.}

\item{type}{Plot either "R2" or "SC" or "Ests" for R2s, structure coefficients
or model estimates.}
}
\description{
Forestplot of the partR2 results
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com),
        Holger Schielzeth  (holger.schielzeth@uni-jena.de)
        Shinichi Nakagawa (s.nakagawa@unsw.edu.au),
}
\keyword{models}
