% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bw}
\alias{get_bw}
\title{Get beta weights.}
\usage{
get_bw(mod, ests)
}
\arguments{
\item{mod}{merMod object.}

\item{ests}{tidy merMod output form broom.mixed}
}
\value{
tidy output with bw instead of raw estimates
}
\description{
For Gaussian models, the predictors and the response are standardised.
For GLMMs, only predictors are standardised.
}
\keyword{internal}
