% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ndf}
\alias{get_ndf}
\title{Get numerator dfs for reduced models}
\usage{
get_ndf(partvar, mod, data)
}
\arguments{
\item{partvar}{One or more fixed effect variables which are taken out
of the model.}

\item{mod}{merMod object.}

\item{data}{Data.frame to fit the model}
}
\value{
Numerator degrees of freedom
}
\description{
Get numerator dfs for reduced models
}
\keyword{internal}
