% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mod_names_partbatch}
\alias{mod_names_partbatch}
\title{Modify term names if partbatch is a named list}
\usage{
mod_names_partbatch(partbatch, part_names)
}
\arguments{
\item{partbatch}{list with batches}

\item{part_names}{character vector with names of terms (and their combinations)}
}
\description{
Modify term names if partbatch is a named list
}
\keyword{internal}
