% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_overdisp}
\alias{model_overdisp}
\title{Adds an observational level random effect to a model}
\usage{
model_overdisp(mod, dat, olre)
}
\arguments{
\item{mod}{merMod object.}

\item{dat}{The underlying data.frame}
}
\description{
Adds an observational level random effect to a model
}
\keyword{internal}
