% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{model_overdisp}
\alias{model_overdisp}
\title{Adds an observational level random effect to a model}
\usage{
model_overdisp(mod, data)
}
\arguments{
\item{mod}{The model.}

\item{data}{The data.frame which has been provided to fit the model in lme4.}
}
\description{
Adds an observational level random effect to a model
}
\keyword{internal}
