% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{model_overdisp}
\alias{model_overdisp}
\title{Adds an observational level random effect to a model}
\usage{
model_overdisp(mod)
}
\arguments{
\item{mod}{The model.}
}
\description{
Adds an observational level random effect to a model
}
\keyword{internal}
