% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partR2-package.R
\docType{package}
\name{partR2-package}
\alias{partR2-package}
\title{partR2: Partitioning R2 in generalized linear mixed models}
\description{
The goals of partR2 are:
}
\details{
\itemize{
\item To estimate marginal and conditional R2 for LMMs and GLMMs.
\item To partition the R2 into variance explained uniquely by each predictor
and variance explained by a combination of predictors using semi-partial R2s.
\item To provide structure coefficients (SC). SC are the correlation between a
predictor and the predicted response, independent of the other predictors.
\item To estimate the the total variance explained by a predictor independent of
other predictors, using inclusive R2 (IR2). IR2 is estimated with SC^2 * R2_full_model.
\item To provide beta weights, which are standardised regression coefficients.
If beta is a model estimate for variable x, and y is the response,then the
beta weight is beta * (sd(x)/sd(y).
\item Use parametric bootstrapping to estimate confidence intervals for
all estimates.
}

The package has one main function \code{\link{partR2}} which takes a fitted model
from lme4. At the moment, Gaussian, Poisson and binomial models are supported.
For Poisson and non-binary binomial models, \code{partR2} adds an
observational level random effect to model additive overdispersion (if
an olre is not fitted already).

The \code{\link{summary.partR2}} function provides an extended summary with R2s, partial
R2s, model estimates and structure coefficients. The \code{\link{forestplot}}
function provides a means of plotting the results.
}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
