% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partR2-package.R, R/partR2.R
\docType{package}
\name{partR2}
\alias{partR2}
\alias{partR2-package}
\title{R2, commonality coefficients and structure coefficients for mixed models}
\usage{
partR2(mod, partvars = NULL, data = NULL, R2_type = "marginal",
  nboot = 10, CI = 0.95, parallel = FALSE, ncpus = NULL)
}
\arguments{
\item{mod}{Gaussian merMod object from lme4::lmer.}

\item{partvars}{Character vector specifying the predictors for which to partition the R2.}

\item{R2_type}{"marginal" or "conditional"}

\item{nboot}{Number of parametric bootstraps for interval estimation
(defaults to NULL). Larger numbers of bootstraps give a better
asymtotic CI, but may be time-consuming. Bootstrapping can be switch on by setting
\code{nboot = 1000}.}

\item{CI}{Width of the required confidence interval between 0 and 1 (defaults to
0.95).}
}
\value{
Returns an object of class \code{partR2} that is a a list with the following elements:
\item{call}{model call}
\item{datatype}{Response distribution (here: 'Gaussian').}
\item{R2_type}{Marginal or conditional R2}
\item{R2_df}{R2 and confidence interval}
\item{R2_boot}{Parametric bootstrap samples for R2}
\item{CC_df}{Commonality coefficients (unique and common R2) and confidence intervals}
\item{SC_df}{Structure coefficients (correlation between predicted response Yhat and the predictors specified in partvars) and confidence intervals}
\item{SC_boot}{Parametric bootstrap samples for the SCs}
\item{partvars}{predictors to partition}
\item{CI}{Coverage of the confidence interval as specified by the \code{CI} argument.}
}
\description{
R2, commonality coefficients and structure coefficients for gaussian lme4 models.
}
\examples{

data(BeetlesBody)
library(lme4)
mod <- lmer(BodyL ~ Sex + Treatment + Habitat + (1|Container) + (1|Population),
            data = BeetlesBody)
(R2 <- partR2(mod,  partvars = c("Treatment", "Sex", "Habitat"), data = BeetlesBody,
                   R2_type = "marginal", nboot = 5, CI = 0.95))
mod <- lmer(BodyL ~ Treatment + Sex + (1 + Treatment|Population),
             data=BeetlesBody)
(R2 <- partR2(mod, partvars = c( "Sex"), data = BeetlesBody,
                   R2_type = "marginal", nboot = 5, CI = 0.95))

data(BeetlesFemale)
mod <- glmer(Egg ~ Treatment + Habitat + (1|Container) + (1|Population),
                  data = BeetlesFemale, family = poisson)
(R2 <- partR2(mod, partvars = c("Treatment", "Habitat"), data = BeetlesFemale,
              R2_type = "marginal", nboot = 5, CI = 0.95))


data(BeetlesMale)

mod <- glmer(Colour ~ Treatment + (1|Container) + (1|Population), data = BeetlesMale,
             family = binomial)
(R2 <- partR2(mod, partvars = c("Treatment"), R2_type = "marginal", nboot = 5, CI = 0.95))

# prepare proportion data // doesn't work yet, as data is difficult to retrieve, don't know yet
how to solve this. (Maybe just add data argument?)
BeetlesMale$Dark <- BeetlesMale$Colour
BeetlesMale$Reddish <- (BeetlesMale$Colour-1)*-1
BeetlesColour <- aggregate(cbind(Dark, Reddish) ~ Treatment + Population + Container,
     data=BeetlesMale, FUN=sum)

mod <- glmer(cbind(Dark, Reddish) ~ Treatment + (1|Container) + (1|Population),
             data = BeetlesMale, family = binomial)
(R2 <- partR2(mod, partvars = c("Treatment"), R2_type = "marginal", nboot = 5, CI = 0.95))

}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.

Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
