% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partR2-package.R, R/partR2.R
\docType{package}
\name{partR2}
\alias{partR2}
\alias{partR2-package}
\title{R2, commonality coefficients and structure coefficients for mixed models}
\usage{
partR2(mod, partvars = NULL, data = NULL, R2_type = "marginal",
  nboot = NULL, CI = 0.95, parallel = FALSE, ncpus = NULL)
}
\arguments{
\item{mod}{merMod object fitted with lme4.}

\item{partvars}{Character vector specifying the predictors for which to partition the R2.}

\item{data}{dataframe used for lme4 model fit.}

\item{R2_type}{"marginal" or "conditional"}

\item{nboot}{Number of parametric bootstraps for interval estimation
(defaults to NULL). Larger numbers of bootstraps give a better
asymtotic CI, but may be time-consuming. Bootstrapping can be switch on by setting
\code{nboot = 1000}.}

\item{CI}{Width of the required confidence interval between 0 and 1 (defaults to
0.95).}

\item{parallel}{If TRUE, computation runs in parallel, leaving one CPU free, except ncpus is specified.}

\item{ncpus}{number of cpus for parallel computation}
}
\value{
Returns an object of class \code{partR2} that is a a list with the following elements:
\item{call}{model call}
\item{datatype}{GLMM Family.}
\item{R2_type}{Marginal or conditional R2}
\item{R2_pe_ci}{R2 and confidence intervals for full model and partitions}
\item{R2_boot}{Parametric bootstrap samples for R2 for full model and partitions}
\item{partvars}{predictors to partition}
\item{CI}{Coverage of the confidence interval as specified by the \code{CI} argument.}
}
\description{
R2, commonality coefficients and structure coefficients for gaussian lme4 models.
}
\examples{

data(BeetlesBody)
library(lme4)

# Gaussian data
mod <- lmer(BodyL ~ Sex + Treatment + Habitat + (1|Container) + (1|Population),
            data = BeetlesBody)
(R2 <- partR2(mod,  data = BeetlesBody, R2_type = "marginal", nboot = 5, CI = 0.95))
(R2 <- partR2(mod,  partvars = c("Treatment", "Sex", "Habitat"), data = BeetlesBody,
                   R2_type = "marginal", nboot = 5, CI = 0.95))

# Random slope (fixed effect Treatment can't be removed separately)
mod <- lmer(BodyL ~ Treatment + Sex + (1 + Treatment|Population),
             data=BeetlesBody)
(R2 <- partR2(mod, partvars = c("Sex"), data = BeetlesBody,
                   R2_type = "marginal"))

# poisson data
data(BeetlesFemale)
mod <- glmer(Egg ~ Treatment + Habitat + (1|Container) + (1|Population),
                  data = BeetlesFemale, family = poisson)
(R2 <- partR2(mod, partvars = c("Treatment", "Habitat"), data = BeetlesFemale,
              R2_type = "marginal", nboot = 5, CI = 0.95))

# binomial data

# binary
data(BeetlesMale)
mod <- glmer(Colour ~ Treatment + (1|Container) + (1|Population), data = BeetlesMale,
             family = binomial)
(R2 <- partR2(mod, partvars = c("Treatment"), R2_type = "marginal", data = BeetlesMale,
             nboot = 5, CI = 0.95))

# proportion
BeetlesMale$Dark <- BeetlesMale$Colour
BeetlesMale$Reddish <- (BeetlesMale$Colour-1)*-1
BeetlesColour <- aggregate(cbind(Dark, Reddish) ~ Treatment + Population + Container,
     data=BeetlesMale, FUN=sum)

mod <- glmer(cbind(Dark, Reddish) ~ Treatment + (1|Container) + (1|Population),
             data = BeetlesColour, family = binomial)
(R2 <- partR2(mod, partvars = c("Treatment"), R2_type = "marginal",  data = BeetlesColour,
             nboot = 5, CI = 0.95))

}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.

Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
