% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partR2-package.R
\docType{package}
\name{partR2}
\alias{partR2}
\alias{partR2-package}
\title{R2, commonality coefficients and structure coefficients for mixed models}
\description{
R2, commonality coefficients and structure coefficients for mixed models
}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
