% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.partR2.R
\name{print.partR2}
\alias{print.partR2}
\title{Print a partR2 object}
\usage{
\method{print}{partR2}(x, round_to = 4, ...)
}
\arguments{
\item{x}{partR2 object returned from one of the partR2 functions}

\item{round_to}{defaults to 4 (decimals)}

\item{\dots}{Additional arguments; none are used in this method.}
}
\value{
No return value, prints concise results of partR2 calculation.
}
\description{
Displays the results a partR2object (i.e. the result of a partR2 function call) in a nice form.
}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed-effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
\keyword{models}
