% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{R2_of_red_mod}
\alias{R2_of_red_mod}
\title{Calculate R2 from a reduced model}
\usage{
R2_of_red_mod(partvar, mod, R2_pe, expct)
}
\arguments{
\item{partvar}{One or more fixed effect variables which are taken out
of the model.}

\item{mod}{merMod object.}

\item{R2_pe}{R2 function.}
}
\value{
R2 of reduced model.
}
\description{
Calculate R2 from a reduced model
}
\keyword{internal}
