% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.partR2.R
\name{summary.partR2}
\alias{summary.partR2}
\title{Complete summary of a partR2 object}
\usage{
\method{summary}{partR2}(object, round_to = 4, ests = FALSE, ...)
}
\arguments{
\item{object}{partR2 object returned from one of the partR2 functions}

\item{round_to}{Defaults to 4 (decimals)}

\item{ests}{Defaults to FALSE, if TRUE, also prints raw model estimates.}

\item{\dots}{Additional arguments; not used at the moment}
}
\value{
No return value, prints extended summary of partR2 calculation.
}
\description{
Displays extended results of partR2, including R2,
part (semi-partial) R2, inclusive R2, structure coefficients and beta weights.
}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed-effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
\keyword{models}
