% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.partR2.R
\name{summary.partR2}
\alias{summary.partR2}
\title{Complete summary of a partR2 object}
\usage{
\method{summary}{partR2}(object, round_to = 4, ...)
}
\arguments{
\item{object}{An partR2 object returned from one of the partR2 functions}

\item{round_to}{defaults to 4 (decimals)}

\item{\dots}{Additional arguments; not used at the moment}
}
\description{
Displays the complete results a partR2object (i.e. the result of a partR2 function call)
which includes R2, partial R2, model estimates and structure coefficients.
}
\references{
Nakagawa, S., & Schielzeth, H. (2013). \emph{A general and simple method for obtaining R2 from
generalized linear mixed‐effects models}. Methods in Ecology and Evolution, 4(2), 133-142.

Newton, R. G., & Spurrell, D. J. (1967).  \emph{A development of multiple regression for the
analysis of routine data. Applied Statistics}. 51-64.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com),
        Shinichi Nakagawa (s.nakagawa@unsw.edu.au),
        Holger Schielzeth  (holger.schielzeth@uni-jena.de)
}
\keyword{models}
