% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_helpers.R
\name{var_comps_poisson}
\alias{var_comps_poisson}
\title{Get variance components for merMod with poisson response.}
\usage{
var_comps_poisson(mod, expct, overdisp_name)
}
\arguments{
\item{mod}{merMod object with poisson family.}

\item{expct}{"meanobs" or "latent". "latent" recommended.}

\item{overdisp_name}{name of overdispersion term}
}
\value{
Fixed, random and residual variance
}
\description{
Get variance components for merMod with poisson response.
}
\keyword{internal}
